tests/wpt/tests/html/canvas/element/pixel-manipulation/2d.imageData.object.properties.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.imageData.object.properties</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.object.properties</h1>
<p class="desc">ImageData objects have the right properties</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData objects have the right properties");
_addTest(function(canvas, ctx) {

  var imgdata = ctx.getImageData(0, 0, 10, 10);
  _assertSame(typeof(imgdata.width), 'number', "typeof(imgdata.width)", "'number'");
  _assertSame(typeof(imgdata.height), 'number', "typeof(imgdata.height)", "'number'");
  _assertSame(typeof(imgdata.data), 'object', "typeof(imgdata.data)", "'object'");

});
</script>
